chore #53 : 안쓰는 파일 제거
