+++ Outline/Outline/Source/Manager/FunctionManager.swift
@@ -43,3 +43,17 @@ func convertToCLLocationCoordinates(_ coordinates: [Coordinate]) -> [CLLocationC
 func convertToCLLocationCoordinate(_ coordinate: Coordinate) -> CLLocationCoordinate2D {
     return CLLocationCoordinate2D(latitude: coordinate.latitude, longitude: coordinate.longitude)
 }
+
+// MARK: - Double Extension 페이스를 표시하기 위한 함수
+extension Double {
+    /// 초단위의 Double 을 페이스 형식으로 바꿔주는 Extension
+    /// - Returns: -'--'' 형식의 String
+    func formattedAveragePace() -> String {
+        if self.isNaN {
+            return "-'--''"
+        }
+        let minutes = Int(self / 60)
+        let seconds = Int(self) % 60
+        return String(format: "%02d'%02d''", minutes, seconds)
+    }
+}
