+++ Outline/Outline/Source/Manager/HealthKitManager.swift
