--- Outline/Outline/Source/View/Running/TimerView.swift
@@ -1,55 +1,0 @@
-//
-//  TimerView.swift
-//  Outline
-//
-//  Created by hyunjun on 10/22/23.
-//
-
-import SwiftUI
-
-struct TimerView: View {
-    var elapsedTime: TimeInterval = 0
-    var showSubseconds: Bool = true
-    @State private var timeFormatter = ElapsedTimeFormatter()
-
-    var body: some View {
-        Text(NSNumber(value: elapsedTime), formatter: timeFormatter)
-            .foregroundColor(Color.primaryColor)
-            .font(.system(size: 40, weight: .bold))
-            .onChange(of: showSubseconds) { value, _ in
-                timeFormatter.showSubseconds = value
-            }
-    }
-}
-
-class ElapsedTimeFormatter: Formatter {
-    let componentsFormatter: DateComponentsFormatter = {
-        let formatter = DateComponentsFormatter()
-        formatter.allowedUnits = [.minute, .second]
-        formatter.zeroFormattingBehavior = .pad
-        return formatter
-    }()
-    var showSubseconds = true
-
-    override func string(for value: Any?) -> String? {
-        guard let time = value as? TimeInterval else {
-            return nil
-        }
-
-        guard let formattedString = componentsFormatter.string(from: time) else {
-            return nil
-        }
-
-        if showSubseconds {
-            let hundredths = Int((time.truncatingRemainder(dividingBy: 1)) * 100)
-            let decimalSeparator = Locale.current.decimalSeparator ?? "."
-            return String(format: "%@%@%0.2d", formattedString, decimalSeparator, hundredths)
-        }
-
-        return formattedString
-    }
-}
-
-#Preview {
-    TimerView()
-}
